Move parse_ver_spend_auth_sig function
